Implemented an extension of the CoreImage class to contain many functions
Allows the code to flow much more ergonomically and really helps with
readability. There are a few more functions that I want to move to this
class, but it nearly stores all logic relating to the manipulation of
these images.

Reworked some function references to reflect these changes too

This was super needed as I start to take on implementing the true
engraving function, which implements the most of these features. In
spirit of folderify and magick, I added a dissolve, which is currently
just a regular composite.